+++ src/States/MapViewState.h
@@ -261,7 +261,8 @@ private:
 	bool				mDebug = false;					/**< Display debug information. */
 	bool				mLeftButtonDown = false;		/**< Used for mouse drags on the mini map. */
 	bool				mLoadingExisting = false;		/**< Flag used for loading an existing game. */
-	bool				mPinResourcePanel = false;		/**< */
+	bool				mPinResourcePanel = false;		/**<  */
+	bool				mPinPopulationPanel = false;	/**<  */
 
 	std::string			mExistingToLoad;				/**< Filename of the existing game to load. */